Satisfy index requests asynchronously
Starts a number of workers (currently hardcoded) that process the
indexing requests.  The job queue is very large (1000), so that even
when full indexing rebuilds arrive, the jobs can be set up and the
server can respond accordingly.

The old queueing code (disp.go) is kept for now.  There will be a big
cleanup of unused functionality sometime later.